add spi_master init() method for resource initialization control
Using .init() instead of .__init__() for resource initialization at python object creation allows finer control
over the initialization time. With .init() a SpiMaster object can be created in a 'parent' process, which passes
the SpiMaster object to a created 'child' process, which in turn initializes the hardware in the 'child' process
execution context.